Completed 4.21

+++ part4/blog_list/controllers/blog.js
@@ -10,6 +10,7 @@ blogRouter.get('/', async (request, response) => {
 })
 
 blogRouter.post('/', async (request, response) => {
+
   const decodedToken = jwt.verify(request.token, process.env.SECRET)
 
   if (!decodedToken.id) {
@@ -42,6 +43,25 @@ blogRouter.post('/', async (request, response) => {
 })
 
 blogRouter.delete('/:id', async (request, response) => {
+
+  const decodedToken = jwt.verify(request.token, process.env.SECRET)
+
+  if (!decodedToken.id) {
+    return response.status(401).json({
+      error: 'token missing or invalid'
+    })
+  }
+
+  const tokenUserId = decodedToken.id
+
+  const blog = await Blog.findById(request.params.id)
+
+  if (blog.user.toString() !== tokenUserId.toString()) {
+    return response.status(401).json({
+      error: 'deletion of blog not authorized'
+    })
+  }  
+
   await Blog.findByIdAndRemove(request.params.id)
   response.status(204).end()
 })
